Use this plugins version of farm.php
There has been a minor change to the original farm.php in lines 52 and
onwards to the end of the block. The idea is that an request to a
non-existing animal is served by the farmer wiki instead of dying with
an error message. This is done in order to be consistent with the
behaviour of a farm implemented with virtual hosts/subdomains.

Known issues: When trying to access a non-existing animal, one is logged
out of the master wiki. This happens because in inc/auth.php auth_logoff
the login-cookie is overwritten.